Rewrite entire codebase in PHP 8.4 style
- Add declare(strict_types=1) to all files
- Add typed properties (string, int, float, bool, union types)
- Add return type declarations to all methods
- Use constructor promotion in Exception, Socket classes
- Use match expressions instead of switch where cleaner
- Use typed class constants (const int, const string)
- Use #[\Override] attribute on overridden methods
- Use #[\Deprecated] attribute on deprecated methods
- Use str_contains()/str_starts_with() instead of strpos/strncmp
- Use null coalescing (??) instead of isset() checks
- Use filter_var() directly (remove extension_loaded checks)
- Use never return type for generateError()
- Use OpenSSLAsymmetricKey type hint in PrivateKey
- Use Shmop type hint in Cache/Shm
- Remove obsolete PHP version checks
- Remove verbose/obvious PHPDoc comments
- Update composer.json: require PHP >=8.4, PHPUnit ^11
- Fix test compatibility with strict types

+++ Net/DNS2/RR/LOC.php
@@ -92,10 +92,10 @@ class Net_DNS2_RR_LOC extends Net_DNS2_RR
 
             $this->longitude = $this->_dms2d($londeg, $lonmin, $lonsec, $lonhem);
 
-            $this->size      = (isset($x[15])) ? $x[15] : 1;
-            $this->horiz_pre = ((isset($x[17])) ? $x[17] : 10000);
-            $this->vert_pre  = ((isset($x[19])) ? $x[19] : 10);
-            $this->altitude  = $x[13];
+            $this->size      = (float)(isset($x[15]) ? $x[15] : 1);
+            $this->horiz_pre = (float)(isset($x[17]) ? $x[17] : 10000);
+            $this->vert_pre  = (float)(isset($x[19]) ? $x[19] : 10);
+            $this->altitude  = (float)$x[13];
 
             return true;
         }
